+++ packages/core/analytics-next/src/index.d.ts
@@ -1,0 +1,3 @@
+// This deliberately empty typing file.
+// There are typings defined in `/typings` folder for internal (AK) usage only.
+declare module '@atlaskit/analytics-next';