+++ packages/media/media-card/__tests__/withDataURI.spec.tsx
@@ -3,6 +3,11 @@ jest.mock('../src/utils/isRetina');
 
 import * as React from 'react';
 import { shallow, ShallowWrapper } from 'enzyme';
+import {
+  MediaItemDetails,
+  DataUri,
+  DataUriService,
+} from '@atlaskit/media-core';
 import { waitUntil } from '@atlaskit/media-test-helpers';
 import {
   withDataURI,
@@ -13,13 +18,27 @@ import {
 import { getElementDimension } from '../src/utils/getElementDimension';
 import { isRetina } from '../src/utils/isRetina';
 
-const DemoComponent = () => null; // tslint:disable-line:variable-name
+interface DemoComponentProps {
+  foo?: string;
+}
+
+class DemoComponent extends React.Component<DemoComponentProps> {
+  render() {
+    return null;
+  }
+}
 const DemoComponentWithDataURI = withDataURI(DemoComponent); // tslint:disable-line:variable-name
 
-const createDataURIService = (imageDataUriDataUri = 'data:jpg') => ({
-  fetchImageDataUri: jest.fn(() => Promise.resolve(imageDataUriDataUri)),
-  fetchOriginalDataUri: jest.fn(),
-});
+const createDataURIService = (
+  imageDataUriDataUri: DataUri = 'data:jpg',
+): jest.Mocked<DataUriService> => {
+  const mock = {
+    fetchOriginalDataUri: jest.fn(),
+    fetchImageDataUri: jest.fn(),
+  } as jest.Mocked<DataUriService>;
+  mock.fetchImageDataUri.mockReturnValue(Promise.resolve(imageDataUriDataUri));
+  return mock;
+};
 
 const waitUntilDataURIIsTruthy = (
   card: ShallowWrapper<WithDataURIProps, WithDataURIState>,
@@ -39,12 +58,15 @@ describe('WithDataURI', () => {
         <DemoComponentWithDataURI />,
       );
 
-      const instance = element.instance() as WithDataURI;
+      const instance = element.instance() as WithDataURI<DemoComponentProps>;
       const updateDataURI = jest.spyOn(instance, 'updateDataURI');
 
-      instance.componentWillReceiveProps({
-        dataURIService: createDataURIService(),
-      });
+      instance.componentWillReceiveProps(
+        {
+          dataURIService: createDataURIService(),
+        },
+        {},
+      );
 
       expect(updateDataURI).toHaveBeenCalledTimes(1);
     });
@@ -54,10 +76,10 @@ describe('WithDataURI', () => {
         <DemoComponentWithDataURI />,
       );
 
-      const instance = element.instance() as WithDataURI;
+      const instance = element.instance() as WithDataURI<DemoComponentProps>;
       const updateDataURI = jest.spyOn(instance, 'updateDataURI');
 
-      instance.componentWillReceiveProps({ metadata: {} });
+      instance.componentWillReceiveProps({ metadata: {} }, {});
 
       expect(updateDataURI).toHaveBeenCalledTimes(1);
     });
@@ -67,9 +89,9 @@ describe('WithDataURI', () => {
         <DemoComponentWithDataURI />,
       );
 
-      const instance = element.instance() as WithDataURI;
+      const instance = element.instance() as WithDataURI<DemoComponentProps>;
       const updateDataURI = jest.spyOn(instance, 'updateDataURI');
-      instance.componentWillReceiveProps({ foo: 'bar' });
+      instance.componentWillReceiveProps({ foo: 'bar' }, {});
 
       expect(updateDataURI).not.toHaveBeenCalled();
     });
@@ -92,7 +114,7 @@ describe('WithDataURI', () => {
 
       element.setState({ dataURI: 'data:png' });
 
-      const instance = element.instance() as WithDataURI;
+      const instance = element.instance() as WithDataURI<DemoComponentProps>;
       instance.updateDataURI({ dataURIService, metadata: undefined });
 
       expect(element.state().dataURI).toBe(undefined);
@@ -101,10 +123,10 @@ describe('WithDataURI', () => {
     it('should clear the dataURI when the metadata is a link', () => {
       const dataURIService = createDataURIService();
 
-      const metadata = {
+      const metadata: MediaItemDetails = {
         url: 'https://example.com',
         title: 'Example link',
-      };
+      } as MediaItemDetails;
 
       const element = shallow<WithDataURIProps, WithDataURIState>(
         <DemoComponentWithDataURI
@@ -114,7 +136,7 @@ describe('WithDataURI', () => {
       );
 
       element.setState({ dataURI: 'data:png' });
-      const instance = element.instance() as WithDataURI;
+      const instance = element.instance() as WithDataURI<DemoComponentProps>;
       instance.updateDataURI({ dataURIService, metadata: undefined });
 
       expect(element.state().dataURI).toBe(undefined);
@@ -133,7 +155,7 @@ describe('WithDataURI', () => {
           metadata={metadata}
         />,
       );
-      const instance = element.instance() as WithDataURI;
+      const instance = element.instance() as WithDataURI<DemoComponentProps>;
       instance.updateDataURI({ dataURIService, metadata });
 
       return waitUntilDataURIIsTruthy(element).then(() =>
@@ -155,7 +177,7 @@ describe('WithDataURI', () => {
         />,
       );
 
-      const instance = element.instance() as WithDataURI;
+      const instance = element.instance() as WithDataURI<DemoComponentProps>;
       instance.updateDataURI({ dataURIService, metadata });
 
       return waitUntilDataURIIsTruthy(element).then(() =>
@@ -177,7 +199,7 @@ describe('WithDataURI', () => {
         { disableLifecycleMethods: true },
       );
 
-      const instance = element.instance() as WithDataURI;
+      const instance = element.instance() as WithDataURI<DemoComponentProps>;
       instance.updateDataURI({ dataURIService, metadata, appearance: 'small' });
 
       expect(
@@ -198,7 +220,7 @@ describe('WithDataURI', () => {
         />,
       );
 
-      const instance = element.instance() as WithDataURI;
+      const instance = element.instance() as WithDataURI<DemoComponentProps>;
       instance.updateDataURI({ dataURIService, metadata, appearance: 'auto' });
 
       expect(
@@ -213,7 +235,7 @@ describe('WithDataURI', () => {
       };
       const invalidDimensions = {
         width: 'hi',
-        height: /a/,
+        height: 'bye',
       };
       const dimensions = {
         width: 50,
@@ -248,7 +270,7 @@ describe('WithDataURI', () => {
           />,
         );
 
-        const instance = element.instance() as WithDataURI;
+        const instance = element.instance() as WithDataURI<DemoComponentProps>;
         instance.updateDataURI({ dataURIService, metadata });
 
         expect(dataURIService.fetchImageDataUri).toBeCalledWith(
@@ -269,7 +291,7 @@ describe('WithDataURI', () => {
           />,
         );
 
-        const instance = element.instance() as WithDataURI;
+        const instance = element.instance() as WithDataURI<DemoComponentProps>;
         instance.updateDataURI({ dataURIService, metadata });
 
         expect(dataURIService.fetchImageDataUri).toBeCalledWith(
@@ -291,7 +313,7 @@ describe('WithDataURI', () => {
           />,
         );
 
-        const instance = element.instance() as WithDataURI;
+        const instance = element.instance() as WithDataURI<DemoComponentProps>;
         instance.updateDataURI({ dataURIService, metadata });
 
         expect(dataURIService.fetchImageDataUri).toBeCalledWith(
@@ -312,7 +334,7 @@ describe('WithDataURI', () => {
           />,
         );
 
-        const instance = element.instance() as WithDataURI;
+        const instance = element.instance() as WithDataURI<DemoComponentProps>;
         instance.updateDataURI({ dataURIService, metadata });
 
         expect(dataURIService.fetchImageDataUri).toBeCalledWith(
@@ -333,7 +355,7 @@ describe('WithDataURI', () => {
           />,
         );
 
-        const instance = element.instance() as WithDataURI;
+        const instance = element.instance() as WithDataURI<DemoComponentProps>;
         instance.updateDataURI({ dataURIService, metadata });
 
         expect(dataURIService.fetchImageDataUri).toBeCalledWith(
@@ -355,7 +377,7 @@ describe('WithDataURI', () => {
           />,
         );
 
-        const instance = element.instance() as WithDataURI;
+        const instance = element.instance() as WithDataURI<DemoComponentProps>;
         instance.updateDataURI({ dataURIService, metadata });
 
         expect(dataURIService.fetchImageDataUri).toBeCalledWith(
@@ -377,7 +399,7 @@ describe('WithDataURI', () => {
         />,
       );
 
-      const instance = element.instance() as WithDataURI;
+      const instance = element.instance() as WithDataURI<DemoComponentProps>;
       instance.updateDataURI({ dataURIService, metadata });
 
       expect(dataURIService.fetchImageDataUri).toHaveBeenCalled();
@@ -393,7 +415,9 @@ describe('WithDataURI', () => {
 
       element.setState({ dataURI: 'data:png' });
 
-      expect(element.find(DemoComponent).props().dataURI).toBe('data:png');
+      expect((element.find(DemoComponent).props() as any).dataURI).toBe(
+        'data:png',
+      );
     });
 
     it('should pass down other props when I am passed them', () => {