+++ packages/editor/conversation/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/conversation
 
+## 6.3.15
+
 ## 6.3.14
 
 ## 6.3.13