+++ packages/editor/editor-confluence-transformer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/editor-cq-transformer
 
+## 2.10.2
+
 ## 2.10.1
 
 ## 2.10.0