+++ packages/media/media-card/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-image
 
+## 25.1.0
+- [minor] Add analytics events for click and show actions of media-card [031d5da](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/031d5da)
+- [minor] Add analytics events for click and show actions of media-card [b361185](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b361185)
+
+## 25.0.2
+
 ## 25.0.1
 
 ## 25.0.0